requirements.txt
gpiozero
rpi-lgpio
flask
mysql-connector-python
flask-socketio
eventlet